Refactor : remove BonusBall class and add bonusBall variable in WinningLotto class
